third_party/blink/web_tests/external/wpt/css/cssom/shorthand-values-expected.txt

This is a testharness.js-based test.
PASS The serialization of border: 1px; border-top: 1px; should be canonical.
PASS The serialization of border: 1px solid red; should be canonical.
PASS The serialization of border: 1px red; should be canonical.
PASS The serialization of border: red; should be canonical.
FAIL The serialization of border-top: 1px; border-right: 1px; border-bottom: 1px; border-left: 1px; should be canonical. assert_equals: expected "border: 1px;" but got "border-width: 1px; border-style: initial; border-color: initial;"
FAIL The serialization of border-top: 1px; border-right: 2px; border-bottom: 3px; border-left: 4px; should be canonical. assert_equals: expected "border-width: 1px 2px 3px 4px;" but got "border-width: 1px 2px 3px 4px; border-style: initial; border-color: initial;"
FAIL The serialization of border: 1px; border-top: 2px; should be canonical. assert_equals: expected "border-width: 2px 1px 1px;" but got "border-width: 2px 1px 1px; border-style: initial; border-color: initial; border-image: initial;"
FAIL The serialization of border: 1px; border-top: 1px !important; should be canonical. assert_equals: expected "border-right-width: 1px; border-bottom-width: 1px; border-left-width: 1px; border-top-width: 1px !important;" but got "border-right: 1px; border-bottom: 1px; border-left: 1px; border-image: initial; border-top: 1px !important;"
FAIL The serialization of border: 1px; border-top-color: red; should be canonical. assert_equals: expected "border-width: 1px; border-top-color: red;" but got "border-width: 1px; border-style: initial; border-right-color: initial; border-bottom-color: initial; border-left-color: initial; border-image: initial; border-top-color: red;"
PASS The serialization of border: solid; border-style: dotted should be canonical.
PASS The serialization of border-width: 1px; should be canonical.
PASS The serialization of overflow-x: scroll; overflow-y: hidden; should be canonical.
PASS The serialization of overflow-x: scroll; overflow-y: scroll; should be canonical.
PASS The serialization of outline-width: 2px; outline-style: dotted; outline-color: blue; should be canonical.
PASS The serialization of margin-top: 1px; margin-right: 2px; margin-bottom: 3px; margin-left: 4px; should be canonical.
PASS The serialization of list-style-type: circle; list-style-position: inside; list-style-image: initial; should be canonical.
PASS The serialization of list-style-type: lower-alpha; should be canonical.
PASS The serialization of font-family: sans-serif; line-height: 2em; font-size: 3em; font-style: italic; font-weight: bold; should be canonical.
PASS The serialization of padding-top: 1px; padding-right: 2px; padding-bottom: 3px; padding-left: 4px; should be canonical.
Harness: the test ran to completion.
